Minor restyling and resizing
Made most of the text and headers smaller because when the site is hosted the font size
increases over what it looks like when developing. Also fixed an issue with the event info
on the home page that keeps the event info intact rather than breaking to new lines when the
window is resized. Also added another sponsor the the events page.