Added gallery images and lightbox
Created a gallery page and populated it with images from the event. Also
added a lightbox plugin to view larger versions of each picture as well as
navigate through the gallery. Still need add a loading screen while the gallery
thumbnails are loading but need to push the current changes through to the live
site per the client request.