Fixed slider, updated content post event and worked on load times
Changed the gallery slider to one that will work with the angular ui router. I also resized the slideshow images to make them more uniform and included them in the preload on the home page. I also updated the landing page per the clients request to reflect on the outcome of the first event. Also made some minor content changes to the events, participate and tickets page since there is not a current event the client is working on. I also included thumbnail pictures for the gallery page to help smooth the scroll top animation when the gallery is revealed.